added block

+++ symfony/recipes/install_pp_au.rb
@@ -18,7 +18,7 @@ node[:deploy].each do |application, deploy|
 
   sites = JSON.parse(file_content)
   sites['site'].each do |current|
-    only_if { ::File.exists?("#{deploy[:deploy_to]}/current/" + current[1]['app_folder'] + "/config/parameters_pp.dist.yml") }
+
 
     #Get vhost and write to file
     tmp_vhost_bucket = vhost_bucket + current[1]['vhost']
@@ -30,6 +30,7 @@ node[:deploy].each do |application, deploy|
       action :create
     end
 
+
     #Get web app file
     obj_app = s3.buckets['ops-works-config'].objects["pp_apps/" + current[1]['web_app']]
     file_content_app = obj_app.read
@@ -38,15 +39,18 @@ node[:deploy].each do |application, deploy|
       action :create
     end
 
+
     #Copy and update parameter file
-    text = File.read("#{deploy[:deploy_to]}/current/" + current[1]['app_folder'] + "/config/parameters_pp.dist.yml")
-    new_contents = text.gsub("YOUR_USER", current[1]['user'])
-    new_contents = new_contents.gsub("YOUR_PASSWORD", current[1]['pass'])
+    ruby_block 'testing' do
+      block do
+        text = File.read("#{deploy[:deploy_to]}/current/" + current[1]['app_folder'] + "/config/parameters_pp.dist.yml")
+        new_contents = text.gsub("YOUR_USER", current[1]['user'])
+        new_contents = new_contents.gsub("YOUR_PASSWORD", current[1]['pass'])
+        # To write changes to the file, use:
+        File.open("#{deploy[:deploy_to]}/current/" + current[1]['app_folder'] + "/config/parameters_pp.yml", "w") {|file| file.puts new_contents }
+      end
+      only_if { ::File.exists?("#{deploy[:deploy_to]}/current/" + current[1]['app_folder'] + "/config/parameters_pp.dist.yml") }
+    end
 
-    # To merely print the contents of the file, use:
-    #puts new_contents
-
-    # To write changes to the file, use:
-    File.open("#{deploy[:deploy_to]}/current/" + current[1]['app_folder'] + "/config/parameters_pp.yml", "w") {|file| file.puts new_contents }
   end
 end
